psiphon: simplify now that we get config from orchestra
Closes #168
