+++ .travis.yml
@@ -3,11 +3,7 @@ go:
 - 1.13.x
 env:
 - CGO_ENABLED=0
-  GO111MODULE=on
 dist: xenial
-before_install:
-- openssl aes-256-cbc -K $encrypted_bfb3eaa23890_key -iv $encrypted_bfb3eaa23890_iv
-  -in testdata/psiphon_config.json.enc -out testdata/psiphon_config.json -d
 before_script:
 - go get golang.org/x/tools/cmd/cover
 - go get github.com/mattn/goveralls
