--- experiment/psiphon/psiphon_otherwise.go
@@ -1,25 +1,0 @@
-// +build nopsiphon
-
-package psiphon
-
-import (
-	"context"
-
-	"github.com/ooni/probe-engine/experiment"
-	"github.com/ooni/probe-engine/experiment/handler"
-	"github.com/ooni/probe-engine/model"
-	"github.com/ooni/probe-engine/session"
-)
-
-// NewExperiment creates a new experiment.
-func NewExperiment(
-	sess *session.Session, config Config,
-) *experiment.Experiment {
-	return experiment.New(
-		sess, testName, testVersion,
-		func(c context.Context, s *session.Session, m *model.Measurement,
-			callbacks handler.Callbacks,
-		) error {
-			return ErrDisabled
-		})
-}
