--- experiment_psiphon_test.go
@@ -1,14 +1,0 @@
-// +build !nopsiphon
-
-package engine
-
-import "testing"
-
-func TestRunPsiphon(t *testing.T) {
-	sess := newSessionForTesting(t)
-	builder, err := sess.NewExperimentBuilder("psiphon")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runexperimentflow(t, builder.Build())
-}
